+++ api/server.go
@@ -440,6 +440,8 @@ func (g *GRPCServer) Start(ctx context.Context, lis net.Listener) error {
 		multiSigSignerEventStore: g.multiSigSignerEventStore,
 		notaryStore:              g.notaryStore,
 		candleServiceV2:          g.candleServiceV2,
+		marketsStore:             g.marketsStore,
+		partiesStore:             g.partyStore,
 	}
 	protoapi2.RegisterTradingDataServiceServer(g.srv, tradingDataSvcV2)