+++ sqlstore/migrations/0001_initial.sql
@@ -362,6 +362,15 @@ create table if not exists markets (
     primary key (id, vega_time)
 );
 
+create view markets_current as (
+    select distinct on (id) id, vega_time, instrument_id, tradable_instrument,
+           decimal_places, fees, opening_auction, price_monitoring_settings,
+           liquidity_monitoring_parameters, trading_mode, state, market_timestamps,
+           position_decimal_places
+    from markets
+    order by id, vega_time desc
+);
+
 CREATE TABLE epochs(
   id           BIGINT                   NOT NULL,
   start_time   TIMESTAMP WITH TIME ZONE NOT NULL,
@@ -443,13 +452,13 @@ CREATE VIEW votes_current AS (
 );
 
 create table if not exists margin_levels (
-    account_id INT  NOT NULL REFERENCES accounts(id),
+    account_id INT  NOT NULL, --REFERENCES accounts(id),
     timestamp timestamp with time zone not null,
     maintenance_margin numeric(32, 0),
     search_level numeric(32, 0),
     initial_margin numeric(32, 0),
     collateral_release_level numeric(32, 0),
-    vega_time timestamp with time zone not null references blocks(vega_time)
+    vega_time timestamp with time zone not null -- references blocks(vega_time)
 );
 
 select create_hypertable('margin_levels', 'vega_time', chunk_time_interval => INTERVAL '1 day');
@@ -686,6 +695,7 @@ DROP TYPE IF EXISTS order_side;
 DROP TYPE IF EXISTS order_type;
 DROP TYPE IF EXISTS order_pegged_reference;
 
+DROP VIEW IF EXISTS markets_current;
 DROP TABLE IF EXISTS markets;
 DROP VIEW IF EXISTS market_data_snapshot;
 DROP TABLE IF EXISTS market_data;